+++ spec/spec_helper.rb
@@ -1,0 +1,29 @@
+require 'pry'
+require 'codeclimate-test-reporter'
+require 'advanced_bot_detection'
+require 'support/constants.rb'
+require 'webmock/rspec'
+
+CodeClimate::TestReporter.start
+
+RSpec.configure do |config|
+  config.expect_with :rspec do |expectations|
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
+  end
+
+  config.mock_with :rspec do |mocks|
+    mocks.verify_partial_doubles = true
+  end
+
+  config.filter_run :focus
+  config.run_all_when_everything_filtered = true
+  config.disable_monkey_patching!
+
+  if config.files_to_run.one?
+    config.default_formatter = 'doc'
+  end
+  config.profile_examples = 10
+  config.order = :random
+  Kernel.srand config.seed
+
+end