+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Version History
 ====
     * All Version bumps are required to update this file as well!!
 ----
-
+* 0.5.0 Updated UA list, actual rspec testing, removed description as something that can come back from a match
+        (this is because we now match multiple types at one time)
 * 0.4.0 Adding in mobile detection features
 * 0.3.0 Forked and renamed from Voight-Kampff gem
 * 0.2.2 Remove active_record requirment, fix google user agent